Add decodeCompressedJson test, Readability changes
